Introduce Interceptor Base protocol
This adds the ability to merge Request and Response interceptors into a single interceptor.
