+++ Sources/Jetworking/Client/Configuration/Configuration.swift
@@ -3,8 +3,7 @@ import Foundation
 /// The configuration used within the client.
 public struct Configuration {
     let baseURL: URL
-    let requestInterceptors: [RequestInterceptor]
-    let responseInterceptors: [ResponseInterceptor]
+    let interceptors: [Interceptor]
     let encoder: JSONEncoder
     let decoder: JSONDecoder
     let requestExecutorType: RequestExecutorType
@@ -12,12 +11,21 @@ public struct Configuration {
     let uploadExecutorType: UploadExecutorType
     let responseQueue: DispatchQueue
 
+    /// RequestInteceptors stored in `interceptors` property
+    public var requestInterceptors: [RequestInterceptor] {
+        interceptors.compactMap { $0 as? RequestInterceptor }
+    }
+
+    /// ResponseInteceptors stored in `interceptors` property
+    public var responseInterceptors: [ResponseInterceptor] {
+        interceptors.compactMap { $0 as? ResponseInterceptor }
+    }
+
     /**
      * Initialises a new configuration instance to use within the client.
      *
      * - Parameter baseURL: The base URL used within the client.
-     * - Parameter requestInterceptors: A list of request interceptors to intercept the request before sending it.
-     * - Parameter responseInterceptors: A list of response interceptors to intercept the response before returning it.
+     * - Parameter interceptors: A list of interceptors to intercept the request before sending (`RequestInterceptor`) it or intersect the response after receiving it (`ResponseInterceptor`).
      * - Parameter encoder: The encoder to use to encode the request body data before sending it.
      * - Parameter decoder: The decoder to use to decode the response body data before returning it.
      * - Parameter requestExecutorType: The request executor type to use to execute the requests.
@@ -26,8 +34,7 @@ public struct Configuration {
      */
     public init(
         baseURL: URL,
-        requestInterceptors: [RequestInterceptor],
-        responseInterceptors: [ResponseInterceptor],
+        interceptors: [Interceptor],
         encoder: JSONEncoder,
         decoder: JSONDecoder,
         requestExecutorType: RequestExecutorType = .async,
@@ -36,8 +43,7 @@ public struct Configuration {
         responseQueue: DispatchQueue = .main
     ) {
         self.baseURL = baseURL
-        self.requestInterceptors = requestInterceptors
-        self.responseInterceptors = responseInterceptors
+        self.interceptors = interceptors
         self.encoder = encoder
         self.decoder = decoder
         self.requestExecutorType = requestExecutorType
