+++ Sources/Jetworking/Client/Interceptor/Protocol/Interceptor.swift
@@ -1,0 +1,4 @@
+import Foundation
+
+/// Base Protocol for `RequestInterceptor` and `ResponseInterceptor`.
+public protocol Interceptor {}
